*** empty log message ***
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@533 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/store.c
@@ -494,7 +494,7 @@ char * object;
 							printf (" %lx", *((char *)(object + z)));
 						}
 #endif
-						buffer_write(object, sizeof(char) * count);
+						buffer_write(object, (int)(sizeof(char) * count));
 
 						break;
 					case SK_FLOAT:
